feat: rebuild => using eslint & confrim one file only have one class  & write a new handleNumClick in Calculator.js
