+++ src/index.css
@@ -1,5 +1,9 @@
-body{
-    padding:10px;
+#root{
+    padding-top: 30px;
+    width: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
 }
 .calculator{
     width: 206px;
