+++ src/utils/Queue.js
@@ -1,0 +1,21 @@
+class Queue{
+  constructor(){
+    this.items = []
+  }
+  enqueue(item){
+    this.items.push(item)
+  }
+  dequeue(){
+    return this.items.shift()
+  }
+  isEmpty(){
+    return this.items.length === 0
+  }
+  size(){
+    return this.items.length
+  }
+}
+
+export {
+  Queue
+}
